week4_library_demo/Backend/auth_service/requirements.txt
flask
flask_sqlalchemy
mysql-connector-python
flask_bcrypt
PyJWT
python-dotenv
gunicorn
Authlib
flask_session